requirements.txt
feedparser
openai
notion-client
python-wordpress-xmlrpc
requests
python-dotenv
schedule
